requirements.txt
prettytable
pymysql
python-openstacksdk
sqlalchemy
